Support more predicates for primitives in queries (#54)
This commit introduces specialized predicates as well some miscellaneous operations for existing primitive types. These include new additions to the String primitive such as `like` and `startsWith`, as well as only allowing the `eq` predicate for the Bool primitive. We do this by introducing a static structure in the predicate builder that maps existing primitives to their supported operations.