Add support for date and time primitives (#112)
This commit adds initial support for date and time primitives, namely `LocalDate`, `LocalTime`, `LocalDateTime`, and `Instant`. The different types encompass different combinations of date, time, and timezone support (check integration tests for specifics). It also introduces the `@precision(n)` annotation for date-time types with time fields, where `n` is the number of digits to keep in the seconds field.