Add support for floats (#137)
This commit adds support for the `Float` primitive, along with support for the `@size` (specifying the minimum size of the field in bytes) and `@bits` hints (specifying the number of bits kept in the float's mantissa). Supported @size values are 4 bytes and 8 bytes, with the supported @bits range being 1 to 53.