Add support for services parsing (#204)
This commit adds parsing for the services spec into the model system as well as introspection of methods and related types. Services are defined using the `service { .. }` block. Service methods are defined using either the `query` and `mutation` keywords, and may be exported using the `export` keyword. A return type must be specified for each method (either a primitive or a composite type).

This commit also adds support for `type { .. }` and `input type { .. }` model definitions, a variant of the existing `model { .. }` blocks. These two models are so-called non-persistent models, and are not meant to be built into the database schema like regular `model`s; instead, they are meant to specify the structure of either input argument types to service methods (`input type`) or the output type of a service method (`type`).

Example usage of both (as well as related annotations) are provided in `integration-tests/services-basic/arithmetic.clay`.