Make 'clay import' work again (#250)
We had set the pool size for import purpose to be 1, but we were passing `Database` to various methods such as `ColumnSpec::from_db`, each of which needed a connection (while another connection was still in use), so the pool size of 1 wasn't sufficient.

So make all those methods take a `Client`. However, `Database.get_client()` return type was too specific and type used in the return type aren't available in payas-cli, so generalize the return type of `get_client()`.