Fix error produced with an empty array and make SQL generated user-data independent
Before this change, we would generate SQL like:
```sql
insert into chess_piece (position_history) values (ARRAY($1, $2))
```
and bind `$1` to the first element and `$2` to the second element, and so on of the use-supplied array in the GraphQL query. This style had two issues.

1. An error when presented with an empty array

For empty arrays, Postgres was not able to decide the type of `Array[]` and suggested that we specify the type like `Array[]::TEXT`. We could have done that, but that would not solve the other issue with the generated SQL.

2. User-data dependent SQL

Here, if the user supplied a 2-element array, we would generate,
```sql
insert into chess_piece (position_history) values (ARRAY($1, $2))
```
whereas with a 3-element array,
```sql
insert into chess_piece (position_history) values (ARRAY($1, $2, $3))
```
this would prevent the possibility of caching SQL statements (where, we would have a `GraphQL Query -> SQL Query` cache, bypassing all the parsing and related logic to create the SQL).

With this commit, we generate the following SQL:
```sql
insert into chess_piece (position_history) values ($1)
```
and bind the whole array to `$1`. This prevents both the issues in one shot.

To deal with multidimensional arrays properly (and keep our abstractions (see QueryContext.literal_value), we had to use the `postgres_array` crate.