Add a rayon thread pool to the test runner in payas-test (#296)
This limits the number of threads that are used when running integration
test files. A server is spawned for every test file, so it is currently
possible to run out of database connections if all the tests are being
run at once - postgres has a default maximum of 100 connections. Even
without this limitation it's probably more efficient to limit the number
of servers spawned, depending on the resources of the test machine.

The number of threads used is the minimum of the number of tests or
double the number of CPUs on the machine.

Also added reporting of time taken and number of CPUs when running the
tests.