Fix corrupt claypot files in integration-tests
Instead of building the claypot file for each production mode test,
which can cause the errors described in #301, we first create a
hashmap of the model paths to the test files which depend on them.

The map is later used to spawn threads which build the claypot file
before running the tests which depend on it once the file has been
built.

This is potentially slower, since multiple tests are running in the
same thread but should at least partially be offset by the reduction
in duplicated `clay build` commands.