Implement one-to-one relationship (#305)
Introduce the concept of cardinality (`Optional` or `Unbounded`) on one-to-many/many-to-one, so that we can reuse most of the implementation. For the `Optional` case, we also add a uniqueness constraint (the only real difference between one-to-many and one-to-one).

We now always look at the other side (for composite types) to check the cardinality and decide the correct cardinality of `OneToMany`/`ManyToOne` relationship.

Also, fix issues when applying filter based on nested relationship.