Remove "dev" tests from integration tests
Running 'clay serve' now builds a claypot file and spawns a clay-server
instance to load it. This means there is no real difference between dev
and production mode since they are both running the same server
executable with a claypot file. The extra time and complexity in the
integration tests is therefore unnecessary.

There are also cleanup problems killing the spawned child server if the
integration tests run 'clay serve' since killing a process doesn't kill its
children.

This means that watching the clay file (for dev mode) will need to be
managed from the cli module, which should kill and restart the spawned
clay-server after rebuilding the claypot file (not done yet).