Add variable binding and dynamic assertions to `payas-test` (#321)
This commit introduces the ability to bind fields from a query to a variable in the `operation` section of integration test files and allows them to be used in the `response` and `variable` sections of test files through the `$` object. In addition, dynamic assertions are now allowed to be specified in `response` by providing a JavaScript closure in lieu of a concrete value. A change has also been made to `DenoModule` in order to allow it to load files from memory. See integration tests for examples.