Allow coercion from reasonable formats for date and time inputs (#326)
Currently, we strictly stick to certain formats for the different date/time fields that we provide; this commit loosens this restriction. Any of our existing formats may be provided to any date/time field given that they have the necessary information (e.g. a date field is allowed to accept anything with a date in it, time fields need to be provided with an input with a time, etc.). We also provide a reasonable error message from `cast_string`.

Fixes #308.