Canonicalize model paths in claytest when passing them to `clay` (#333)
This seems to be necessary when working with a model file that is a directory above the actual `.claytest` file, otherwise `clay` will fail to find the model file. This commit also exposes some errors from the parser; previously we were reporting a generic `Error: Failed to parse input file` when it could have been IO-related or a missing file.