Implement a cleaner way to compute joins (#340)
Earlier we used to do an ad hoc computation of the joins needed by looking at the predicate of a query alone. We did this by analyzing reachability of columns used in the predicate through the root type. This worked in simple cases, but not in many others: multiple fields of the same type, columns with an indirection (concert -> concert_artists -> artist).

This change computes the table dependency when computing the predicate itself.

* Fix a typo in `PredicateParameterTypeKind::Operator`
* Reduce duplication when building logical operator predicates
* Rename `boolean_param` to `logical_op_param`
* Gather column dependencies when computing predicates
* Introduce the concept of `ColumnPath` and use it to compute predicate along with any needed joins
* Integrate column_paths to compute joins
* Add a test to show new capability of deep joins
* Update integration test with an example of logical expression with nested predicate