Remove the --watch flag from clay serve and build
In both cases, it had no effect. For `clay serve`, we watch
the parent directory of the model file automatically. For `clay build`,
we didn't watch it anyway.

Given that `clay serve` is a dev tool, developers will expect it to
watch automatically (and if not they can always call
`clay build && clay-server`). For `clay build`, it is a build
tool (most likely used as a part of a deployment script), so watching
is unnecessary.

Also changed the message that we watch the parent directory instead
of just the model file.