Move response stream creation to the core payas-server crate
The actix crate now just depends on futures::stream rather than
the async_stream library.

The resolve function in the core payas-server module now returns
a stream which can contain either a successful JSON response or
JSON containing information on error(s) that occurred.

This means that the only public functionality required is the
create_system_info function (and its opaque return type) and the
resolve function which returns an opaque 'impl Stream'.