Split payas-server into payas-server-core and payas-server-actix (#385)
The the code with actix dependencies is now in payas-server-actix and
calls into the payas-server-core crate which contains most of the resolver-related
code.

SystemInfo is now an opaque type since it is only really used internally
in payas-server-core. Server implementations just need to be able to create
and store an instance, passing it to the core `resolve` function.

Removed some unused code in payas-server which became apparent after
making the modules private.

The resolve function in the payas-server-core crate now returns
a stream which can contain either a successful JSON response or
JSON containing information on error(s) that occurred.

This means that the only public functionality required is the
create_system_info function (and its opaque return type) and the
resolve function which returns an opaque 'impl Stream'.