Introduce a separate and robust validation stage (#387)
This change introduces a separate validation stage, which verifies the query's adherence to the schema, resolution of all variables, and resolutions of all fragments. The output of the validation process is a full resolved and normalized query. For example, the output query expands fragments to be resolved. As a result, the data resolution becomes much simpler.

The changes to integration tests reflects some of the incorrect queries we have been accepting.

Note that this commit does not address validating variable types in the validation stage (it continues to happen during the resolution phase).

Fixes #130
Fixes #309
Fixes #312

Also:
- fix an issue snapshot naming with a newer version of insta
- update insta version in payas-test to match the rest
- move the insta dependency in payas-server-core to dev-dependencies