Properly report query parse errors
We had an `unwrap()` upon query parsing. Replace that with a proper error reporting. Some of the errors (such as duplicate fragments) report two positions, so expand `ExecutionError` to accomodate upto two positions. We could have generalized it to report any number of positions, but currently there are no use cases for more than two positions, and it will incur an allocation (not the biggest concern, but a consideration, nevertheless!).

Attempted to create a macro for position reporting, but couldn't get it to compile.
```rust
macro_rules! report_position {
    ($position:tt) => {
        yield Bytes::from_static(br#"{"line": "#);
        yield Bytes::from($position.line.to_string());
        yield Bytes::from_static(br#", "column": "#);
        yield Bytes::from($position.column.to_string());
        yield Bytes::from_static(br#"}"#);
    };
}
```