Properly report query parse errors (#400)
We had an `unwrap()` upon query parsing. Replace that with a proper error reporting. Some of the errors (such as duplicate fragments) report two positions, so expand `ExecutionError` to accommodate up to two positions. We could have generalized it to report any number of positions, but currently there are no use cases for more than two positions, and it will incur an allocation (not the biggest concern, but a consideration, nevertheless!).

Co-authored-by: Andy Chun <andy.chun@payalabs.com>