Add tracing instrumentation (#403)
* Add tracing instrumentation

Two out of the box options are provided. See the  README and
telemetry.rs for more information.

Tracing instrumentation locations in the code are kind of arbitrary
for now and will have to be tuned as we work out how best to use it.

Similarly how the telemetry pipeline is tuned for efficiency (when
outputting data to a system like Jaeger) is something that will be
researched as required.