Ignore __typename while validating arguments (#410)
A few typical usages of GraphQL operations involve taking an old value (typically has the `__typename` attribute while querying--often added by clients such as Apollo for caching purposes), update that value, and then send it as an argument for an update mutation. To support such cases this commit makes `argument_validator` not consider the `__typename` argument as a stray argument.