Support commonly found postgres URLs
This issue was encountered when trying to integrate with YugabyteDB, which issues the URL in the
form of:
postgresql://<user>:<password>@<host>:<port>/<db>?ssl=true&sslmode=verify-full&sslrootcert=<crt-file-path>

The tokio-postgres library does not support these ssl params. While we could have asked users to
simplify this URL and supply additional params such as `sslrootcert` as an env var, it is something
we have to explain and could be confusing.

This PR cleans up the URL and configures the SSL connectors according to those parameters.