Filter only requested fields from a response (#447)
We had this filtering logic only for Deno-resolved services, but that didn't cover the case where an around interceptor returned an object with more fields that requested. So this change moves the logic to a higher level of processing the response to cover the original case as well as the around interceptor case.