Implement privileged claytip (#457)
Currently, services must operate with the same context as the invoking user. However, there are models that must not be accessed directly by users, yet need access by services. For example,
- Login flow: The `User` model should not be queried/mutated by a non-admin users, but service methods such as `register` or `login` need to access the same model.
- RSVP: A non-admin user should be able to create an RSVP (perhaps after passing a captcha test), but not able to even query it.

This change introduces a concept of "privileged" `Claytip` that offers a new method: `executeQueryPriv`. This method takes an extra argument to override any context. Access rules are evaluated with this new context (delegating to the underlying context if there is no override for a particular field).