Propagate explicitly thrown errors through interceptors and executeQuery (#477)
Previously, we let any explicitly thrown error become a generic one resulting in a generic "Internal Server Error" response. This change fixes the issue by detecting explicitly thrown errors and wrapping them in a custom error object. 

To detect deeply nested errors, we need to get to the last error in the error chain. However, our use of `#[error(transparent)]` did not allow that to happen (the source method in such case returns the source of the `#[from]` marked field). So we change all `#[error(transparent)]` with `#[error("{0}")]` which delegate the display to the underlying error, yet gives us access to the source (which we use recursively to get to
the last error in the chain).

Fixes #426