Simplify RequestContext implementation (#493)
Remove test/non-test config. Instead, introduce `TestRequestContext`
bound to the "test" annotation. This context may be given a JSON value
to use as the config. Also, introduces the "test-context" feature in
payas-parser so that we process the `@test` context only during tests.

This simplifies the code a bit and (more importantly) uses the same code
path to resolve a context value for both test and non-test code.

Also fixes a couple of typos found in the code.