Remove validation logic's dependency on schema (#501)
Earlier, we used to use the `Schema` object (which also served another purpose--serving introspection). However, the schema was pretty lossy (for example, didn't include a reference to the original type). As a result, validation was more "light" (did check for the object shapes, but only basic checks when it comes to leaf primitive types).

Here, we use traits to normalize `GqlType`, `Query` etc. Thus we don't need the actual `Schema` object for validation.

Note that in this PR, we still use `Schema` for validation the introspection queries themselves. Also, we don't do the deep validation yet.