Add support for IP addresses in request contexts (`@clientIp`) (#511)
This commit adds support for the `@clientIp` request context source, allowing users to extract the originating IP addresses from requests. It includes a rewrite of the AWS Lambda server to allow usage of the full event metadata provided to the Lambda function, and also adds a basic unit test to test the server locally.