Add a Deno prelude section to claytest files and support async checks (#334)
This commit adds support for the `deno:` section in `.claytest` files. The contents of this section will be injected directly into the start of the `assert.js` script, allowing the user to import modules for custom checks or to declare custom functions in. We also introduce the ability for the user to specify asynchronous functions as custom checks. See integration tests for usage.

The commit also adds `TypescriptLoader`, a Deno module loader that is able to load and transpile TypeScript modules from URLs at runtime. This is necessary to be able to use URL imports in `clay test`. The loader can be enabled or disabled at compilation time using the `typescript-loader` feature of `payas-deno` (it is currently on by default).