Pluginify into subsystems (#529)
This is a huge PR, because there was no reasonable way to break it into multiple (we did some of that in earlier PRs towards this goal).

Modularise building and resolution into a plugin-based system, where collaboration between the core system and each plugin in kept to a minimal. We now have a few traits: `SubsystemBuilder`, `SubsystemLoader`, and `SubsystemResolver` that ensures that the model for each subsystem remains opaque to the core claytip code. Each `SubsystemBuilder`s builds a subsystem model and serialize its content to a `Vec<u8>`; the other side `SubsystemLoader` does the opposite and create a `SubsystemResolver` with that subsystem model. All such subsystem resolvers form the overall `SystemResolver` that delegates resolving as appropriate.

Along the way, we treat interceptor in a similar way.