Drop the "payas" prefix from most crates (#535)
With all the renaming and rearranging in recent PRs, this is a good time to drop the "payas" prefix from most core crates. We considered replacing it with "claytip", but that feels redundant given the repository name.

This PR following  rust-analyzer structure to create "crates" and "libs". The former contain code that belongs to this repo in the long run, whereas the one in "libs" may be separated out (for those we kept the "payas" prefix, since we don't know under what name we will fork those).

All changes are mechanical in nature. Except "payas-test" is named "testing" (the word "test" conflicts with `#[test]`.