Rename "database" to "postgres" (#539)
This better reflects the fact that we support the Postgres database,
makes a room for other database plugins, and is more consistent with
how we named the "deno" plugin (and not the "js" plugin).

When we support other databases, we will probably refactor any common
code into a separate "database_util" module (the same way we have shared
code between the "deno" and "wasm" plugins).

We leave "payas-sql" unchanged, since it has a lot more
database-agnostic code (all the "asql" stuff, for example). When we
will support other databases, we will have to revisit its internal
structure and/or making multiple crates out of it.