Weave correct interceptor even with multiple interceptors of the same name
Earlier, we used the interceptor name as the key in the interceptor map.
This meant that if there were multiple interceptors with the same name,
an arbitrary one would be used. This commit fixes this by using the service
name along with the interceptor name as the key in the map.

Fixes #526