Make top-level syntax plugin-agnostic (#568)
Services must now be explicitly tagged with the subsystem they are targeting. The `@external` annotation has been eliminated.

Previously, this was inferred by what kind of model was declared; in particular, `model`s were implicitly assumed to be `postgres` database schemas and services were implied to  either `deno` or `wasm` (depending on the extension in the `@external` annotation). See integration tests for details.

A number of other changes have been made as well:
- `SubsystemBuilder`s can now declare as a part of their trait implementation what annotations they support
- `SubsystemBuilder`s are now dynamically discovered (with a directory search for libraries named `*_model_builder`)
- The VSCode extension was updated for the new syntax changes