Support 'type' in all places we support 'model' (#581)
This is part one of the overall change from using `type` in place of `model`. This one retains the capability of using either `model` or `type`. In a separate PR, we will remove the use of `model`.

One of the major change here is that we had a bug where every plugin process every type. So if a clay file had a deno and wasm services, both the demo and wasm plugin would have processed each others types. We didn't catch this so far, since the postgres plugin processes only `model`s and deno/wasm plugin processed `type`s and we didn't have a test with both deno and wasm.