Remove the `Service` variant of `Type` (#590)
Instead, create a top-level `Service` type. This is a bit clearer and
prevents creation of invalid system (for example have a field of `Service`).

Also, introduce a `TypecheckedSystem` type, which encapsulates types
and services of a system. This is a step towards further cleanup (for
example, have only contexts at the top-level).