Fix return type modifiers for postgres mutations (#600)
We were returning optional types all for mutations, which is not correct
for the creation mutations. We also returned list of optional types for
all bulk mutations.

Now we return non-optional types for creation mutations
and list of non-optional types for all bulk mutations.

Filed https://github.com/payalabs/payas/issues/599 to fix this properly
(the behavior will remain the same, but the code will be cleaner).