Accept id in data param for update mutation
A typical way clients use update mutation is to get the data along with the id,
modify the data and send it back to the server. So we should accept the id
as an optional field in the update mutation to avoid clients having to
separate the id from the data.

See also https://github.com/payalabs/payas/issues/601