Support aggregate queries (#605)
With this change, we now support aggregate queries such as `count`, `max`, `sum`, `avg`. Specifically, for each user type, we support a new top-level query: `<pluralized-lower-case-user-type>Agg` such as `concertsAgg` and `venuesAgg` as well as a nested aggregate query for each field with one-to-many relationship. For example, we support a `concertsAgg` field in the `venue`/`venues` query.

Each of these queries support a `where` argument with the same meaning as the one for a normal query.