Let annotations precede the field they annotate
Before this change, we let annotations precede all other elements (service, types, queries, mutations, and interceptors). So the only place we deviated from this arrangement is for fields. This change fixes that. This also makes the grammar more consistent with the way typescript supports annotations.

To make it clear what is being annotated, we also allow an optional semicolon after a field declaration.

The real change is in grammar.js . The rest is just updating the tests to match the new syntax.