Allow annotations precede the field they annotate (#607)
Before this change, we let annotations precede all other elements (service, types, queries, mutations, and interceptors). But we deviated from this arrangement for fields. This change fixes that. This also makes the grammar more consistent with the way typescript supports annotations.

To make it clear what is being annotated, we also allow an optional semicolon after a field declaration.

The real change is in grammar.js . The rest is just updating the tests to match the new syntax.