Separate primitive and composite type (#617)
This is the first PR to cleanup the Postgres plugin.

The focus here is to properly construct type-safe one-to-many and many-to-one relationships. Before the change, we had `PostgresType` to include both primitives and composite types. In a relation, we used such `PostgresType` and thus could construct a relation between a composite and a primitive or even between two primitives.

The same structure also forced us to check (and bail out) that a certain type is composite. Even though we had been careful during the build phase that such check would not fail, it is error prone and checking at runtime is best avoided.

This change fixes that by keep primitives separate and relation refer to only composite types.