Separate entity and mutation types
Earlier, we had a single type for both entities and mutations:
`PostgresCompositeType`. This made the code unnecessarily complex,
and increased the possibility of looking into the wrong arena.

This commit separates the two types, and simplifies the `MutationType`.
The `MutationType` now only contains the fields that are relevant for
mutations (for example, no need to carry the `plural_name` around).