Support recursive types in a more general way through `FieldType`
Earlier, each usage had its own version of supporting recursive types to support list/optional decoration. We now introduce `FieldType` that can decorate a base type with list/optional decoration in a recursive manner. Using this support:

(Note: we had a `FieldType` in the postgres plugin, which is now renamed to `PostgresFieldType`)

* Generalize OperationReturnType to support any recursive modifications and move it to core
* Use decorated types for: ServiceFieldType, ResolvedFieldType (in service and postgres plugin)