Cleanup postgres type_builder
Instead of creating EntityType in each expansion function, we mutate
the existing entity. This is more efficient, requires less code, clearly
express what is being expanded, and also allows us to avoid cloning
fields from the existing entity.

Also, rename "model_type" to "entity_type" to be consistent with the
rest of the codebase.