Upgrade deno to 1.31
The biggest change is that in 1.30, deno no longer exposes the `Deno.core`
object. Instead, we need to use internal `Deno[Deno.internal].core` object.
Furthermore, the way to call ops is also changed. We need to use
``Deno[Deno.internal].core.ops.<the-name-of-the-op-in-the-rust-code>`,
where such an op is defined using `#[op]` macro.

Also change the ops names to be more consistent with the `op_<ClassName>_<MethodName>`
convention. For example, `op_add_header` is changed to `op_claytip_add_header`.