Upgrade deno to 1.31 (#633)
The biggest change is that in 1.30, deno no longer exposes the `Deno.core`
object [1]. Instead, we need to use an internal `Deno[Deno.internal].core` object.
Furthermore, the way to call ops is also changed. We need to use
`Deno[Deno.internal].core.ops.<the-name-of-the-op-in-the-rust-code>`,
where such an op is defined using `#[op]` macro [2] (instead of Deno.core.opSync("<the-name-of-the-op-in-the-rust-code>" or Deno.core.opAsync("<the-name-of-the-op-in-the-rust-code>").

Also change the ops names to be more consistent with the `op_<ClassName>_<MethodName>`
convention. For example, `op_add_header` is changed to `op_claytip_add_header`.

[1] https://github.com/denoland/deno/pull/16881
[2] https://discord.com/channels/684898665143206084/778060818046386176/1069324824511528980