Rename "clay serve" to "clay dev" (#641)
The "serve" part didn't convey the meaning (that it will run in the
development mode with watching and schema checking). Furthermore,
due to "clay-server" which is production mode, it caused confusion.

So, we now have "clay dev" for development mode and "clay-server" for
production mode. It also matches how node.js ecosystem uses the "dev"
subcommand (such as "npm run dev", "yarn dev", or "next run dev").