Document and refactor access related code in the core module (#645)
* Document access related code in the core module
* Fix typing of `AccessContextSelection` and `AccessSolver`

Earlier, `AccessContextSelection` which is supposed to represent a path from the context object allowed to just have the context object itself. We fix this by representing the path as a non-empty list.

`AccessSolver::extract_context` now returns a `Map<String, Value>` instead of a generic json `Value`. This is because the context object cannot be a list or a scalar value, etc.

Co-authored-by: Luke Taylor <tekul@pm.me>