Fix nested predicates in delete mutation (#651)
We support nested mutations such as in the GraphQL schema:
```graphql
deleteConcerts(where: {venue: {name: {eq: "v1"}}}) {
   ...
}
```
However, executing such an operation generated an invalid SQL, which caused the operation to fail. The main reason was that we produced SQL of the following kind:
```sql
delete from concerts where venue.name = "v1"
```
which is obviously incorrect (where does the `venues` table come from).

Note that the same kind of predicate works fine for queries since there we perform a join to bring in the `venues` relationship. But the join idea won't work for delete mutation, since a `delete from`  cannot be given a join.

We now produce a subselect. For example, in the GraphQL mutation above, we generate the following kind of SQL:
```sql
delete from concerts where concerts.venue_id in (select id from venues where venues.name = "v1"
```

During the implementation, we encountered a lifetime issue, where we were making certain lifetimes too restrictive (for example, the `AbstractSelect` argument in select_transformer). So we removed those lifetimes. This further forced us to refactor the `Literal` variants of `Column` and `ColumnPath` to hold onto an `Arc`. And that itself led to the need for `SqlParamContainer`, since `Arc` is not a `#[fundamental]` type (see comments in `SqlParamContainer`).

Also added tests for predicate (and moved the logic of mapping an `AbstractPredicate` to `Predicate` to predicate_tranformer in line with other transformers).

Of course, we also added integration tests for nested predicates in delete.