Remove the the lifetime parameter on PredicateTransformer.to_subselect_predicate()
This required a few other changes (some of which are not strictly), but
the main change includes removing the MaybeOwned wrapper around the
predicate's constituent parts. Instead we make the predicate's constituent
parts owned by the predicate itself. Since usually we are dealing with
`Column`s `ColumnPath`s, this is not a big deal (they wrap around references
to `PhysicalColumn` and such).