Remove the the lifetime parameter on PredicateTransformer.to_subselect_predicate()
This change is needed to implement the next fix (duplicated entries when supplying the where clauses with a one-to-many field). However, since we had to make big changes already and these changes have a standing on their own, making this a PR.

This required a few other changes (some of which are not strictly), but the main change includes removing the `MaybeOwned` wrapper around the predicate's constituent parts. Instead we make the predicate's constituent
parts owned by the predicate itself. Since usually we are dealing with `Column`s `ColumnPath`s, this is not a big deal (they wrap around references to `PhysicalColumn` and such).

Also remove the `MaybeOwned` wrappers around `Column`s json variants and `Select`'s preciates

Also, introduce the `GroupBy` primitive.