Disallow ordering on a one-to-many field (#660)
It doesn't make sense to oder by an one-to-many field. For example, it is ill-formed to order venues
by the concerts' id (assuming venue -> [concert] relationship) or order concerts by artists' name
(assuming concert -> [artist] relationship). In each case, the ordering is ambiguous (since there
are many artists per concert, and many concerts per venue).

This fixes the generation of order by types to not include fields that are one-to-many
relationships.