Statically link the Postgres and Deno resolver
The use of dynamic loading for Postgres and Deno plugin made it difficult to produce tracing statement (would need the initialization logic in each plugin, with unclear consequences) and it made starting server slower (from 1ms to 10-20ms). See #580 for more background, though that issue suggests a different approach, which we might still pursue later.

With this change, we statically link Postgres and Deno plugins. We still keep the capability to dynamically load these plugins (through feature flags) as well as load other resolver plugins dynamically.