Fix invalid schema diff with uniqueness constraints (#668)
We used a `Vec` to store column names in `CreateUniqueConstraint`. This meant if the order in that `Vec` is swapped  (which can happen depending on how we read information from the database), we would declare it as a mismatch during migrations and schema verification.

We now use a `HashSet` to correctly model columns in a uniqueness constraint.