Fix rust-analyzer code navigation in postgres-model
We had a dev-dependency on `postgres-model` in `core-resolver`, which
caused rust-analyzer to think we have a circular dependency between
them (this is a bug in rust-analyzer, see comments in
document_validator.rs for more details). This was causing issues with
code navigation in Postgres subsystem.

Here we replace dev-dependency in `core-resolver` from `postgres-model`
to `core-plugin-interface`. This does create a different circular
dependency, but it's a lot more local (core-* code only) and makes
life easier with Postgres subsystem.