Fix SQL generation for one-to-many predicate along with a nested order by (#673)
The SQL generation logic didn't work when user supplied a predicate with a one-to-many and an order-by that refers to a nested column. This change fixes that. For example, if we have `[Notification] - Concert - [ConcertArtist] - Artist`, we can support predicate based on any of the notifications' direct/indirect fields along with any order by such as the following query:

```graphql
query getNotifications($search: String) {
  notifications(
    where: {or: [{title: {ilike: $search}}, {concert: {or: [{title: {ilike: $search}}, {concertArtists: {artist: {name: {ilike: $search}}}}]}}]}
    orderBy: {concert: {title: ASC}}
  ) {
  ...
  }
}
```

In the process, we fixed how `select_transform` works. Specifically, we have a concept of `SelectionStrategy` that can examine the query and decides if it is suitable for that query. We also introduce a `SelectionStrategyChain` that looks for a suitable strategy and then invokes its transformation.

Also, this separates low-level selection function from the top-level one.