Add init file to basic-model-no-auth/mutations tests
This file creates an extra venue and concert which are now used in
the concerts-update.claytest file. Some other tests in the mutations
folder also needed expected venue and concert ids incremented to
accomodate this change.

This is really a test of the test system itself, to make sure that
extra init files are run after the one in the parent folder.