Support subdirectories in integration tests (#680)
If a model file is not found in the same directory as the test file,
parent directories will be checked until one is found (until the root
test directory is reached).

Mutations from basic-model-no-auth were moved into the 'mutations'
subfolder to test this works. This folder also now contains an init file
which creates an extra venue and concert. These are used in the
concerts-update.claytest file. Some other tests in the mutations
folder also needed expected venue and concert ids incremented to
accommodate this change (this is really a test of the test system itself,
to make sure that extra init files are run after the one in the parent folder).