Support more forms of access control for mutations (#683)
So far, we have supported access control that evaluates to `Predicate::True`/`Predicate::False`, which meant users cannot use `self` to express fine-grained rules. This change fixes that for update and delete mutations. For create mutations, we still support rules that don't use `self`, but now we enforce that (this is being tracked in #69).

During testing, we encountered that `predicate_transformer` doesn't deal with `param op column` (in situations such as`AuthContext.id == self.user.id`), so fixed that as well.

Also, rename `ColumnPath::Literal` to `ColumnPath::Param`.

Fixes #343 #180