Restructrue CLI to colocate commands and execution
So far, structure for all commands and code to extract argument
was defined in the main.rs file. That made is hard to find the
correspondance between the command and the code that executes it.

This commit introduces the `CommandDefinition` trait that has a method
to return the structure and another to execute the command. This
encapsulates both aspects into a single place.