Fix a dependency between testing and resolver (#707)
This was a source of intermittent build failures (we have experienced it in CI a few times as well
as during Docker build on certain machines). The `testing` subproject relied on the `graphql` node
package in `graphiql` being installed, but those dependencies are fetched only when `resolver` is
build (see its build.rs). If it happens that `testing` gets built before `resolver` (which can
happen since `testing` doesn't (and need not) have `resolver` as a dependency), the build would
fail.

This commit fixes the issue by making the `testing` subproject bring its own `graphql` node package.