Use ephemeral database for integration testing
This PR lifts the implementation of local unix-socket based postgres
(with fallback to Docker) used in the "yo" command to make it more
generally available for integration testing.

Also introduces some abstraction to separate creating a database
server and creating a database. This allows integration-tests to
use the same server (but different databases it is) for all tests.

This gets rid of the need to specify the `EXO_TEST_POSTGRES_URL`.
Updated the README to reflect this.